Introduzindo o Maven no código
