+++ backend/src/main/java/com/wegone/ManualDeOperacao.java
@@ -1,3 +1,4 @@
+package backend.src.main.java.com.wegone;
 import java.util.Date;
 import java.util.Scanner;
