+++ backend/src/main/java/com/wegone/translate/en.json
@@ -1,0 +1,4 @@
+{
+    "welcome": "Welcome",
+    "exit": "Sair"
+}
