+++ backend/src/main/java/com/wegone/translate/pt.json
@@ -1,0 +1,26 @@
+{
+    "welcome": "Bem-vindo ao WegOne",
+    "choice-type-manual": "Escolha o tipo do manual",
+    "maintenence-manual": "Manual de manutenção",
+    "operation-manual": " Manual de operação",
+    "security-manual": "Manual de segurança",
+    "operation-conduct-manual": "Manual de conduta operacional",
+    "diagnostic-manual": "Maual de diagnostico",
+    "new-manual-sucessefully-added": "Novo manual inserido com sucesso",
+    "choice-the-type-of-manual-to-view": "Escolha o tipo de manual para visualizar:",
+    "existing-manuals": "Manuais existentes:",
+    "invalid-option": "Opção inválida!",
+    "select-a-manual-to-see-more-details": "Selecione um manual para ver mais detalhes: ",
+    "choose-the-type-of-manual-you-want-to-delete": "Escolha o tipo de manual que deseja apagar:",
+    "choose-the-number-of-the-manual-you-want-to-delete": "Escolha o número do manual que deseja apagar:",
+    "manual-sucessfully-deleted": "Manual apagado com sucesso!",
+    "choose-the-type-of-the-manual-you-want-to-edit": "Escolha o tipo de manual que deseja editar:",
+    "choose-the-number-of-the-manual-you-want-to-edit": "Escolha o número do manual que deseja editar;",
+    "what-would-you-want-to-edit": "O que você deseja editar?",
+    "title": "Título",
+    "author": "Autor",
+    "text": "Texto",
+    "publication-date": "Data de publicação",
+    "closing-the-program": "Encerrando o programa...",
+    "try-again": "Opção inválida. Tente novamente!"
+}
